get request delete file

+++ excelapp/views.py
@@ -48,8 +48,6 @@ class GenerateContracts(generics.CreateAPIView):
         # Fayllarni va direktoriyalarni aniqlash
         base_dir = Path(__file__).parent if "__file__" in locals() else Path.cwd()
         word_template_path = base_dir / "amaliyot11.docx"
-        if f'papka_{request.user.id}':
-            shutil.rmtree(f'papka_{request.user.id}')
 
         output_dir = base_dir / f"papka_{request.user.id}"  # yuklanadigan papka
         upload_dir = base_dir / "UPLOAD_excel"  # UPLOAD katalogi
@@ -67,7 +65,8 @@ class GenerateContracts(generics.CreateAPIView):
             doc.render(record)
             output_path = output_dir / f"{record['Talabaning_F_I_Sh']}-amaliyot.docx"
             doc.save(output_path)
-        os.remove(saved_file_path)
+        if saved_file_path:
+            os.remove(saved_file_path)
         # shutil.rmtree(output_dir)
 
         return Response({'success': 'Fayl yuborildi', 'yuklanadigan papka joyi': str(output_dir)},
@@ -163,5 +162,5 @@ class DownloadOutputView(APIView):
 
         # Remove the ZIP file after preparing the response
         os.remove(zip_file_path)
-        # shutil.rmtree(output_dir)
+        shutil.rmtree(output_dir)
         return response
